+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [5.12.0] - 2025-07-08
+## [5.12.0] - 2025-07-17
+- added severity, service and resource type to operational rules report
+- changed operational overview report to new format
+- changed operational resources to correspond to overview
 - added `GET /resources` and `GET /resources/arn/{arn}` endpoints to retrieve resources with optional filtering and pagination.
 - added `ResourceCollector` class for retrieving resources using Cloud Custodian filters.